requirements.txt
yagmail==0.15.293
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
filetype==1.2.0
duckduckgo-search
lxml
exa-py
